refactor packages; increase delta.io package version in compatibility sbt config
