+++ build.sbt
@@ -346,7 +346,7 @@ lazy val compatibility = (project in file("oss-compatibility-tests"))
       // Test Dependencies
       "org.scalatest" %% "scalatest" % "3.0.5" % "test",
       "org.apache.spark" % "spark-sql_2.12" % "3.0.0" % "test",
-      "io.delta" % "delta-core_2.12" % "0.8.0" % "test",
+      "io.delta" % "delta-core_2.12" % "1.0.0" % "test",
       "commons-io" % "commons-io" % "2.8.0" % "test",
       "org.apache.spark" % "spark-catalyst_2.12" % "3.0.0" % "test" classifier "tests",
       "org.apache.spark" % "spark-core_2.12" % "3.0.0" % "test" classifier "tests",
